refactor(title-block-zen): rename scss files to use .module.scss
