+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/MobileActions.tsx
@@ -1,8 +1,8 @@
+import React from "react"
+import classnames from "classnames"
 import { Icon } from "@kaizen/component-library"
 import { ButtonProps } from "@kaizen/button"
 import { MenuItem, MenuList } from "@kaizen/draft-menu"
-import classnames from "classnames"
-import * as React from "react"
 import chevronDownIcon from "@kaizen/component-library/icons/chevron-down.icon.svg"
 import chevronUpIcon from "@kaizen/component-library/icons/chevron-up.icon.svg"
 import {
@@ -14,7 +14,7 @@ import {
   TitleBlockMenuItemProps,
   convertSecondaryActionsToMenuItems,
 } from "./TitleBlockZen"
-import styles from "./MobileActions.scss"
+import styles from "./MobileActions.module.scss"
 
 const buttonIsLink: (action: TitleBlockButtonProps) => boolean = action =>
   "href" in action
