+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/NavigationTabs.module.scss
@@ -4,7 +4,7 @@
 @import "~@kaizen/deprecated-component-library-helpers/styles/type";
 @import "~@kaizen/component-library/styles/fonts";
 @import "~@kaizen/deprecated-component-library-helpers/styles/layout";
-@import "./TitleBlockBreakpoints.scss";
+@import "./mixins";
 
 .container {
   display: flex;
