+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/SecondaryActions.tsx
@@ -1,12 +1,11 @@
+import React from "react"
 import { Button, IconButton } from "@kaizen/button"
 import { Menu, MenuList, MenuItem } from "@kaizen/draft-menu"
-import * as React from "react"
 import chevronDownIcon from "@kaizen/component-library/icons/chevron-down.icon.svg"
 import meatballsIcon from "@kaizen/component-library/icons/meatballs.icon.svg"
 import { SecondaryActionsProps, TitleBlockMenuItemProps } from "./TitleBlockZen"
 import Toolbar from "./Toolbar"
-
-import styles from "./TitleBlockZen.scss"
+import styles from "./TitleBlockZen.module.scss"
 
 type Props = {
   secondaryActions?: SecondaryActionsProps
