+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/_variables.scss
@@ -1,4 +1,5 @@
 @import "~@kaizen/design-tokens/sass/color";
+
 $dt-color-background-color-admin: $color-gray-100;
 $dt-color-background-color-eduction: $color-blue-100;
 $dt-color-background-color-default: $color-purple-600;
