+++ draft-packages/title-block-zen/docs/TitleBlockZen.stories.module.scss
@@ -1,0 +1,3 @@
+.contentContainer {
+  background: #eee;
+}
