--- draft-packages/title-block-zen/docs/TitleBlockZen.stories.scss
@@ -1,10 +1,0 @@
-@import "~@kaizen/deprecated-component-library-helpers/styles/layout";
-@import "../KaizenDraft/TitleBlockZen/TitleBlockBreakpoints.scss";
-
-body {
-  margin: 0;
-}
-
-.contentContainer {
-  background: #eee;
-}
